(operation) extract ExpotentialBackoff function
