+++ src/delegation_backend/aws_keyspaces.go
@@ -17,34 +17,6 @@ import (
 	logging "github.com/ipfs/go-log/v2"
 )
 
-// Operation is a function type that represents an operation that might fail and need a retry.
-type Operation func() error
-
-const (
-	maxRetries     = 10
-	initialBackoff = 1 * time.Second
-)
-
-// ExponentialBackoff retries the provided operation with an exponential backoff strategy.
-func ExponentialBackoff(operation Operation, maxRetries int, initialBackoff time.Duration) error {
-	backoff := initialBackoff
-	var err error
-	for i := 0; i < maxRetries; i++ {
-		err = operation()
-		if err == nil {
-			return nil // Success
-		}
-
-		if i < maxRetries-1 {
-			// If not the last retry, wait for a bit
-			time.Sleep(backoff)
-			backoff *= 2 // Exponential increase
-		}
-	}
-
-	return fmt.Errorf("operation failed after %d retries, returned error: %s", maxRetries, err)
-}
-
 // InitializeKeyspaceSession creates a new gocql session for Amazon Keyspaces using the provided configuration.
 func InitializeKeyspaceSession(config *AwsKeyspacesConfig) (*gocql.Session, error) {
 	auth := sigv4.NewAwsAuthenticator()
